+++ functions.h
@@ -1,0 +1,9 @@
+#ifndef FUNCTIONS_H_INCLUDED
+#define FUNCTIONS_H_INCLUDED
+
+void writeData();
+void displayData();
+void editTask(int id);
+void completeTask(int id);    // This functions takes ID number as a argument and deletes that task from note list
+
+#endif // FUNCTIONS_H_INCLUDED
