+++ include/TaskClass.h
@@ -1,0 +1,16 @@
+#ifndef TASKCLASS_H
+#define TASKCLASS_H
+
+
+class TaskClass
+{
+    public:
+        TaskClass();
+        virtual ~TaskClass();
+
+    protected:
+
+    private:
+};
+
+#endif // TASKCLASS_H
