+++ src/TaskClass.cpp
@@ -1,0 +1,11 @@
+#include "TaskClass.h"
+
+TaskClass::TaskClass()
+{
+    //ctor
+}
+
+TaskClass::~TaskClass()
+{
+    //dtor
+}
